[update] 网盟entity修改

+++ perfect-entity/src/main/java/com/perfect/entity/report/NmsAdReportEntity.java
@@ -77,7 +77,7 @@ public class NmsAdReportEntity extends AccountIdEntity {
     private Double hopRate;    // 二跳率
 
     @Field(value = "art")
-    private Long avgResTime;  // 平均访问时间
+    private Double avgResTime;  // 平均访问时间
 
     @Field(value = "dt")
     private Integer directTrans;    // 直接转化
@@ -254,11 +254,11 @@ public class NmsAdReportEntity extends AccountIdEntity {
         this.hopRate = hopRate;
     }
 
-    public Long getAvgResTime() {
+    public Double getAvgResTime() {
         return avgResTime;
     }
 
-    public void setAvgResTime(Long avgResTime) {
+    public void setAvgResTime(Double avgResTime) {
         this.avgResTime = avgResTime;
     }
