+++ perfect-entity/src/main/java/com/perfect/entity/report/NmsGroupReportEntity.java
@@ -67,7 +67,7 @@ public class NmsGroupReportEntity extends AccountIdEntity {
     private Double hopRate;    // 二跳率
 
     @Field(value = "art")
-    private Long avgResTime;  // 平均访问时间
+    private Double avgResTime;  // 平均访问时间
 
     @Field(value = "dt")
     private Integer directTrans;    // 直接转化
@@ -219,11 +219,11 @@ public class NmsGroupReportEntity extends AccountIdEntity {
         this.hopRate = hopRate;
     }
 
-    public Long getAvgResTime() {
+    public Double getAvgResTime() {
         return avgResTime;
     }
 
-    public void setAvgResTime(Long avgResTime) {
+    public void setAvgResTime(Double avgResTime) {
         this.avgResTime = avgResTime;
     }
